Fixed deleteUserController: now when a user is deleted, its reference is also removed from his friends' friends lists

+++ controllers/User.js
@@ -10,12 +10,12 @@ export const addUserController = async (req, res) => {
         if (Object.keys(newUser).length === 0) {
             serverResponse(res, 400, "Bad request");
         }
-        const emailRegex = new RegExp(/^[A-Za-z0-9_!#$%&'*+\/=?`{|}~^.-]+@[A-Za-z0-9.-]+$/, "gm");
+        const emailRegex = new RegExp(/^[A-Za-z0-9_!#$%&'*+/=?`{|}~^.-]+@[A-Za-z0-9.-]+$/, "gm");
         if (!emailRegex.test(newUser.email)) {
-            return serverResponse(res, 400, { message: "The email is not in valid form"});
+            return serverResponse(res, 400, { message: "The email is not in valid form" });
         }
         if (!enforceStrongPassword(req.body.password)) {
-            return serverResponse(res, 400, { message: "The password is too weak, it should be no fewer than 8 characters, with at least one uppercase letter, at least one lowercase letter, at least one digit and at least one special character"});
+            return serverResponse(res, 400, { message: "The password is too weak, it should be no fewer than 8 characters, with at least one uppercase letter, at least one lowercase letter, at least one digit and at least one special character" });
         }
         newUser.password = await hashPassword(req.body.password);
         const user = await addUser(newUser);
@@ -24,7 +24,7 @@ export const addUserController = async (req, res) => {
         console.log(e);
         serverResponse(res, 500, e);
     }
-    
+
 }
 
 export const getUserByIdController = async (req, res) => {
@@ -40,16 +40,40 @@ export const getUserByIdController = async (req, res) => {
             message: "Internal error while trying to get user details"
         });
     }
+}
 
+export const getAllFriendsByUserId = async (req, res) => {
+    try {
+        const id = req.params.id;
+        const user = await getUserById(id);
+        if (!user) {
+            return serverResponse(res, 404, { message: "User doesn't exist" });
+        }
+        return serverResponse(res, 200, user.friends);
+    } catch (e) {
+        return serverResponse(res, 500, {
+            message: "Internal error while trying to get user's friends"
+        });
+    }
 }
 
 export const deleteUserController = async (req, res) => {
     try {
         const id = req.params.id;
-        const deletedUser = await deleteUserById(id);
-        if (!deletedUser) {
-            return serverResponse(res, 404, { message: "User doesn't exist"});
+        const userToDelete = await getUserById(id);
+        if (!userToDelete) {
+            return serverResponse(res, 404, { message: "User doesn't exist" });
         }
+        const friendsOfUserToDelete = [...userToDelete.friends];
+        for (const friendId of friendsOfUserToDelete) {
+            const friendToDeleteUserFrom = await getUserById(friendId);
+            const newFriendsList = [...friendToDeleteUserFrom.friends];
+            const ind = newFriendsList.findIndex(friendId => friendId.toString() === id);
+            newFriendsList.splice(ind, 1);
+            friendToDeleteUserFrom.friends = newFriendsList;
+            await friendToDeleteUserFrom.save();
+        }
+        const deletedUser = await deleteUserById(id);
         return serverResponse(res, 200, deletedUser);
     } catch (e) {
         return serverResponse(res, 500, {
@@ -66,7 +90,7 @@ export const updateUserController = async (req, res) => {
     }
     try {
         const id = req.params.id;
-        const user = await getUserById({ _id: id });
+        const user = await getUserById(id);
         if (!user) {
             return serverResponse(res, 404, { message: `Invalid updates - user with id ${id} doesn't exist` });
         }
@@ -89,12 +113,12 @@ export const updateUserPasswordController = async (req, res) => {
     }
     try {
         const id = req.params.id;
-        const user = await getUserById({ _id: id });
+        const user = await getUserById(id);
         if (!user) {
             return serverResponse(res, 404, { message: `Invalid updates - user with id ${id} doesn't exist` });
         }
         if (!enforceStrongPassword(req.body.password)) {
-            return serverResponse(res, 400, { message: "The password is too weak, it should be no fewer than 8 characters, with at least one uppercase letter, at least one lowercase letter, at least one digit and at least one special character"});
+            return serverResponse(res, 400, { message: "The password is too weak, it should be no fewer than 8 characters, with at least one uppercase letter, at least one lowercase letter, at least one digit and at least one special character" });
         }
         user.password = await hashPassword(req.body.password);
         user.updatedAt = Date.now();
@@ -109,19 +133,20 @@ export const updateUserPasswordController = async (req, res) => {
 
 export const addFriendController = async (req, res) => {
     try {
-        const friends = {...req.body};
-        const user = await getUserById({_id: friends.userId});
-        const newFriend = await getUserById({_id: friends.friendId});
+        const friends = { ...req.body };
+        const user = await getUserById(friends.userId);
+        const newFriend = await getUserById(friends.friendId);
         if (!user || !newFriend) {
             return serverResponse(res, 404, { message: `Either the user or the new friend doesn't exist` });
         }
         user.friends.push(newFriend);
+        newFriend.friends.push(user);
         await user.save();
+        await newFriend.save();
         return serverResponse(res, 200, user);
     } catch (e) {
         return serverResponse(res, 500, {
             message: "Internal error while trying to add a user to the friends list"
         });
     }
-
 }
